requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlalchemy>=2.0
mysql-connector-python>=8.4
pydantic>=2.7
alembic>=1.13
python-dotenv>=1.0
